Factor out common channel functionality into base class
