+++ src/channels/session.ts
@@ -1,23 +1,17 @@
-import { WebSocket } from 'ws';
-import { Socket, type Channel } from 'phoenix';
+import { type Channel } from 'phoenix';
 import { SevereServiceError } from 'webdriverio';
-import { Logger } from '../logger.js';
-import { getServiceInfo } from '../utils.js';
+import { BaseChannel } from './base.js';
 
 import type {
   TVLabsCapabilities,
-  TVLabsSocketParams,
   TVLabsSessionRequestEventHandler,
   TVLabsSessionRequestResponse,
   LogLevel,
 } from '../types.js';
-import type { PhoenixChannelJoinResponse } from '../phoenix.js';
 
-export class SessionChannel {
-  private socket: Socket;
+export class SessionChannel extends BaseChannel {
   private lobbyTopic: Channel;
   private requestTopic?: Channel;
-  private log: Logger;
 
   private readonly events = {
     SESSION_READY: 'session:ready',
@@ -29,22 +23,18 @@ export class SessionChannel {
   } as const;
 
   constructor(
-    private endpoint: string,
-    private maxReconnectRetries: number,
-    private key: string,
-    private logLevel: LogLevel = 'info',
+    endpoint: string,
+    maxReconnectRetries: number,
+    key: string,
+    logLevel: LogLevel = 'info',
   ) {
-    this.log = new Logger('@tvlabs/session-channel', this.logLevel);
-    this.socket = new Socket(this.endpoint, {
-      transport: WebSocket,
-      params: this.params(),
-      reconnectAfterMs: this.reconnectAfterMs.bind(this),
-    });
-
-    this.socket.onError((...args) =>
-      SessionChannel.logSocketError(this.log, ...args),
+    super(
+      endpoint,
+      maxReconnectRetries,
+      key,
+      logLevel,
+      '@tvlabs/session-channel',
     );
-
     this.lobbyTopic = this.socket.channel('requests:lobby');
   }
 
@@ -193,79 +183,7 @@ export class SessionChannel {
     }
   }
 
-  private async join(topic: Channel): Promise<void> {
-    return new Promise((res, rej) => {
-      topic
-        .join()
-        .receive('ok', (_resp: PhoenixChannelJoinResponse) => {
-          res();
-        })
-        .receive('error', ({ response }: PhoenixChannelJoinResponse) => {
-          rej('Failed to join topic: ' + response);
-        })
-        .receive('timeout', () => {
-          rej('timeout');
-        });
-    });
-  }
-
-  private async push<T>(
-    topic: Channel,
-    event: string,
-    payload: object,
-  ): Promise<T> {
-    return new Promise((res, rej) => {
-      topic
-        .push(event, payload)
-        .receive('ok', (msg: T) => {
-          res(msg);
-        })
-        .receive('error', (reason: string) => {
-          rej(reason);
-        })
-        .receive('timeout', () => {
-          rej('timeout');
-        });
-    });
-  }
-
-  private params(): TVLabsSocketParams {
-    const serviceInfo = getServiceInfo();
-
-    this.log.debug('Info:', serviceInfo);
-
-    return {
-      ...serviceInfo,
-      api_key: this.key,
-    };
-  }
-
-  private reconnectAfterMs(tries: number) {
-    if (tries > this.maxReconnectRetries) {
-      throw new SevereServiceError(
-        'Could not connect to TV Labs, please check your connection.',
-      );
-    }
-
-    const wait = [0, 1000, 3000, 5000][tries] || 10000;
-
-    this.log.info(
-      `[${tries}/${this.maxReconnectRetries}] Waiting ${wait}ms before re-attempting to connect...`,
-    );
-
-    return wait;
-  }
-
   private tvlabsSessionLink(sessionId: string) {
     return `https://tvlabs.ai/app/sessions/${sessionId}`;
-  }
-
-  private static logSocketError(
-    log: Logger,
-    event: ErrorEvent,
-    _transport: new (endpoint: string) => object,
-    _establishedConnections: number,
-  ) {
-    log.error('Socket error:', event.error);
   }
 }
